agc040/aを追加

+++ Library/rle.cpp
@@ -3,10 +3,10 @@ using namespace std;
 
 /* encode: ランレングス圧縮を行う
 */
-vector<pair<char, int>> encode(const string& str) {
-    int n = (int)str.size();
-    vector<pair<char, int>> ret;
-    for (int l = 0; l < n;) {
+vector<pair<char, long long>> encode(const string& str) {
+    int n = (long long)str.size();
+    vector<pair<char, long long>> ret;
+    for (long long l = 0; l < n;) {
         int r = l + 1;
         for (; r < n && str[l] == str[r]; r++) {};
         ret.push_back({str[l], r - l});
@@ -16,10 +16,10 @@ vector<pair<char, int>> encode(const string& str) {
 }
 /* decode: ランレングス圧縮の復元を行う
 */
-string decode(const vector<pair<char, int>>& code) {
+string decode(const vector<pair<char, long long>>& code) {
     string ret = "";
     for (auto p : code) {
-        for (int i = 0; i < p.second; i++) {
+        for (long long i = 0; i < p.second; i++) {
             ret.push_back(p.first);
         }
     }
